wolfram we can request podindex because it messes up some results
query affected was "3 months from now"
from the api doc:
If you are using the podtimeout parameter, it is important to note that the index numbers are decided at the close of the "scan" stage. The culling of pods based on index is done before the later format stage, so pods that are dropped because they time out in formatting have no bearing on the index numbers.